+++ core-ng/src/main/java/core/framework/internal/log/appender/KafkaAppender.java
@@ -53,7 +53,7 @@ public final class KafkaAppender implements LogAppender {
             Map<String, Object> config = Map.of(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri,
                     ProducerConfig.ACKS_CONFIG, "0",                                    // no acknowledge to maximize performance
                     ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis(),  // metadata update timeout
-                    ProducerConfig.COMPRESSION_TYPE_CONFIG, CompressionType.SNAPPY.name,
+                    ProducerConfig.COMPRESSION_TYPE_CONFIG, CompressionType.ZSTD.name,
                     ProducerConfig.LINGER_MS_CONFIG, 50,
                     ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, (int) Duration.ofSeconds(60).toMillis(),     // DELIVERY_TIMEOUT_MS_CONFIG is INT type
                     ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder");      // if not specify, kafka uses producer-${seq} name, also impact jmx naming