+++ core-ng/src/test/java/core/framework/internal/kafka/MessageListenerTest.java
@@ -13,12 +13,17 @@ class MessageListenerTest {
 
     @BeforeEach
     void createMessageListener() {
-        listener = new MessageListener(null, null, null);
+        listener = new MessageListener("localhost:9092", null, null);
     }
 
     @Test
     void listenerThreadName() {
         assertThat(listener.listenerThreadName(null, 0)).isEqualTo("kafka-listener-0");
         assertThat(listener.listenerThreadName("name", 2)).isEqualTo("kafka-listener-name-2");
+    }
+
+    @Test
+    void consumer() {
+        assertThat(listener.consumer()).isNotNull();
     }
 }