+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
-### 6.10.4 (11/29/2018 - )
+### 6.10.4 (11/29/2018 - ) !!! kafka producer compression changes only works with kafka 2.1.0+
+* kafka: update COMPRESSION_TYPE_CONFIG to ZSTD, requires kafka 2.1.0 on server side
 
 ### 6.10.3 (11/24/2018 - 11/29/2018)
 * search: update es to 6.5.1